+++ src/solvers/dgsem_p4est/dg_3d_gpu.jl
@@ -5,6 +5,12 @@
 @muladd begin
 #! format: noindent
 
+@inline flux_differencing_workgroupsize(::HalfSweep, ::Val) = ()
+@inline function flux_differencing_workgroupsize(::FullSweep,
+                                                 ::Val{NNODES}) where {NNODES}
+    return ((NNODES, NNODES, NNODES, 1),)
+end
+
 @inline function calc_volume_integral!(backend::Backend, du, u,
                                        mesh::Union{P4estMesh{3}, T8codeMesh{3}},
                                        have_nonconservative_terms, equations,
@@ -12,15 +18,19 @@
                                        dg::DGSEM, cache)
     @unpack derivative_split = dg.basis
     @unpack contravariant_vectors = cache.elements
-    kernel! = flux_differencing_KAkernel!(backend)
     NNODES = nnodes(dg)
+    kernel_type = flux_differencing_kernel_type(cache)
+    kernel! = flux_differencing_KAkernel!(backend,
+                                          flux_differencing_workgroupsize(kernel_type,
+                                                                          Val(NNODES))...)
     kernel!(du, u, equations,
             typeof(mesh),
+            kernel_type,
             have_nonconservative_terms,
             combine_conservative_and_nonconservative_fluxes(volume_integral.volume_flux,
                                                             equations),
             dg,
-            volume_integral.volume_flux, Val(NNODES),
+            volume_integral.volume_flux, Val(NNODES), Val(nvariables(equations)),
             derivative_split,
             contravariant_vectors,
             ndrange = (NNODES, NNODES, NNODES, nelements(dg, cache)))
@@ -30,14 +40,16 @@ end
 @kernel function flux_differencing_KAkernel!(du, u, equations,
                                              MeshT::Type{<:Union{P4estMesh{3},
                                                                  T8codeMesh{3}}},
+                                             ::HalfSweep,
                                              have_nonconservative_terms::False,
                                              combine_conservative_and_nonconservative_fluxes::False,
                                              dg::DGSEM,
                                              volume_flux,
                                              ::Val{NNODES},
+                                             ::Val{NVARIABLES},
                                              derivative_split,
                                              contravariant_vectors,
-                                             alpha = true) where {NNODES}
+                                             alpha = true) where {NNODES, NVARIABLES}
     # `true * [some floating point value] == [exactly the same floating point value]`
     # This can (hopefully) be optimized away due to constant propagation.
     i, j, k, element = @index(Global, NTuple)
@@ -138,14 +150,16 @@ end
 @kernel function flux_differencing_KAkernel!(du, u, equations,
                                              MeshT::Type{<:Union{P4estMesh{3},
                                                                  T8codeMesh{3}}},
+                                             ::HalfSweep,
                                              have_nonconservative_terms::True,
                                              combine_conservative_and_nonconservative_fluxes::True,
                                              dg::DGSEM,
                                              volume_flux,
                                              ::Val{NNODES},
+                                             ::Val{NVARIABLES},
                                              derivative_split,
                                              contravariant_vectors,
-                                             alpha = true) where {NNODES}
+                                             alpha = true) where {NNODES, NVARIABLES}
     # `true * [some floating point value] == [exactly the same floating point value]`
     # This can (hopefully) be optimized away due to constant propagation.
     i, j, k, element = @index(Global, NTuple)
@@ -244,6 +258,142 @@ end
                                            fluxtilde3_right,
                                            i, j, kk, element)
     end
+end
+
+@kernel function flux_differencing_KAkernel!(du, u, equations,
+                                             MeshT::Type{<:Union{P4estMesh{3},
+                                                                 T8codeMesh{3}}},
+                                             ::FullSweep,
+                                             have_nonconservative_terms::False,
+                                             combine_conservative_and_nonconservative_fluxes::False,
+                                             dg::DGSEM,
+                                             volume_flux,
+                                             ::Val{NNODES},
+                                             ::Val{NVARIABLES},
+                                             derivative_split,
+                                             contravariant_vectors,
+                                             alpha = true) where {NNODES,
+                                                                  NVARIABLES}
+    # `true * [some floating point value] == [exactly the same floating point value]`
+    # This can (hopefully) be optimized away due to constant propagation.
+    i, j, k, element = @index(Global, NTuple)
+
+    u_local = @localmem eltype(du) (NVARIABLES, NNODES, NNODES, NNODES)
+
+    u_node_global = get_node_vars(u, equations, dg, i, j, k, element)
+    @inbounds for v in 1:NVARIABLES
+        u_local[v, i, j, k] = u_node_global[v]
+    end
+    @synchronize
+
+    u_node = get_node_flux(u_local, Val(NVARIABLES), i, j, k)
+    du_local = zero(SVector{NVARIABLES, eltype(du)})
+
+    Ja1_node = get_contravariant_vector(1, contravariant_vectors, i, j, k, element)
+    for ii in 1:NNODES
+        # pull the contravariant vectors and compute the average
+        Ja1_avg = 0.5f0 * (Ja1_node +
+                   get_contravariant_vector(1, contravariant_vectors,
+                                            ii, j, k, element))
+        # compute the contravariant volume flux in the direction of the
+        # averaged contravariant vector
+        fluxtilde1 = volume_flux(u_node,
+                                 get_node_flux(u_local, Val(NVARIABLES), ii, j, k),
+                                 Ja1_avg, equations)
+        du_local = du_local + (alpha * derivative_split[i, ii]) * fluxtilde1
+    end
+
+    Ja2_node = get_contravariant_vector(2, contravariant_vectors, i, j, k, element)
+    for jj in 1:NNODES
+        Ja2_avg = 0.5f0 * (Ja2_node +
+                   get_contravariant_vector(2, contravariant_vectors,
+                                            i, jj, k, element))
+        fluxtilde2 = volume_flux(u_node,
+                                 get_node_flux(u_local, Val(NVARIABLES), i, jj, k),
+                                 Ja2_avg, equations)
+        du_local = du_local + (alpha * derivative_split[j, jj]) * fluxtilde2
+    end
+
+    Ja3_node = get_contravariant_vector(3, contravariant_vectors, i, j, k, element)
+    for kk in 1:NNODES
+        Ja3_avg = 0.5f0 * (Ja3_node +
+                   get_contravariant_vector(3, contravariant_vectors,
+                                            i, j, kk, element))
+        fluxtilde3 = volume_flux(u_node,
+                                 get_node_flux(u_local, Val(NVARIABLES), i, j, kk),
+                                 Ja3_avg, equations)
+        du_local = du_local + (alpha * derivative_split[k, kk]) * fluxtilde3
+    end
+
+    add_to_node_vars!(du, du_local, equations, dg, i, j, k, element)
+end
+
+@kernel function flux_differencing_KAkernel!(du, u, equations,
+                                             MeshT::Type{<:Union{P4estMesh{3},
+                                                                 T8codeMesh{3}}},
+                                             ::FullSweep,
+                                             have_nonconservative_terms::True,
+                                             combine_conservative_and_nonconservative_fluxes::True,
+                                             dg::DGSEM,
+                                             volume_flux,
+                                             ::Val{NNODES},
+                                             ::Val{NVARIABLES},
+                                             derivative_split,
+                                             contravariant_vectors,
+                                             alpha = true) where {NNODES,
+                                                                  NVARIABLES}
+    # `true * [some floating point value] == [exactly the same floating point value]`
+    # This can (hopefully) be optimized away due to constant propagation.
+    i, j, k, element = @index(Global, NTuple)
+
+    u_local = @localmem eltype(du) (NVARIABLES, NNODES, NNODES, NNODES)
+
+    u_node_global = get_node_vars(u, equations, dg, i, j, k, element)
+    @inbounds for v in 1:NVARIABLES
+        u_local[v, i, j, k] = u_node_global[v]
+    end
+    @synchronize
+
+    u_node = get_node_flux(u_local, Val(NVARIABLES), i, j, k)
+    du_local = zero(SVector{NVARIABLES, eltype(du)})
+
+    Ja1_node = get_contravariant_vector(1, contravariant_vectors, i, j, k, element)
+    for ii in 1:NNODES
+        Ja1_avg = 0.5f0 * (Ja1_node +
+                   get_contravariant_vector(1, contravariant_vectors,
+                                            ii, j, k, element))
+        fluxtilde1_left, _ = volume_flux(u_node,
+                                         get_node_flux(u_local, Val(NVARIABLES),
+                                                       ii, j, k),
+                                         Ja1_avg, equations)
+        du_local = du_local + (alpha * derivative_split[i, ii]) * fluxtilde1_left
+    end
+
+    Ja2_node = get_contravariant_vector(2, contravariant_vectors, i, j, k, element)
+    for jj in 1:NNODES
+        Ja2_avg = 0.5f0 * (Ja2_node +
+                   get_contravariant_vector(2, contravariant_vectors,
+                                            i, jj, k, element))
+        fluxtilde2_left, _ = volume_flux(u_node,
+                                         get_node_flux(u_local, Val(NVARIABLES),
+                                                       i, jj, k),
+                                         Ja2_avg, equations)
+        du_local = du_local + (alpha * derivative_split[j, jj]) * fluxtilde2_left
+    end
+
+    Ja3_node = get_contravariant_vector(3, contravariant_vectors, i, j, k, element)
+    for kk in 1:NNODES
+        Ja3_avg = 0.5f0 * (Ja3_node +
+                   get_contravariant_vector(3, contravariant_vectors,
+                                            i, j, kk, element))
+        fluxtilde3_left, _ = volume_flux(u_node,
+                                         get_node_flux(u_local, Val(NVARIABLES),
+                                                       i, j, kk),
+                                         Ja3_avg, equations)
+        du_local = du_local + (alpha * derivative_split[k, kk]) * fluxtilde3_left
+    end
+
+    add_to_node_vars!(du, du_local, equations, dg, i, j, k, element)
 end
 
 function prolong2interfaces_and_calc_interface_flux!(backend::Backend,